refactor(Vsix): match namespace to "Json" folder.
